Move swagger bootstrapping to http api helper
